requirements.txt
# Core dependencies
google-generativeai==0.8.3
requests==2.32.3
python-dotenv==1.0.1
pydantic>=2.11.0,<3.0.0

# Web UI
streamlit==1.41.1
plotly==5.24.1
pandas==2.2.3

# Development & Testing
pytest==8.3.4
pytest-cov==6.0.0

# Code quality
ruff==0.8.4
